Wrap functions with completion! macro
Completion records as defined in the spec either have a state of normal where the value is returned or have a state of throw where the record itself is returned to caller. See spec: https://timothygu.me/es-howto/#completion-records-and-shorthands

We now wrap any functions that deal with completion records with the completion! macro that is a shorthand for either unwrapping the value or returning a throw completion.